[expo-cli] Move cmd.exe detection from 'bin/expo' to 'src/windows' file
